refactor(source): move schema registry to parent mod
